refactor: 重构传输系统为统一资源适配器模式，优化缓存与性能
- **重构传输系统为统一资源适配器模式**：引入 `LogisticsResource<C>` 接口，将物品、流体、能量、化学品、魔源、热量、魔力等所有传输类型统一接入标准管线。删除原有的 `TransferType` 记录、`*TransferHandler` 和 `*Transfer` 类，新增 `ResourceAdapterHandler` 和 `ResourceAdapterProtocol` 作为通用适配器，通过 `TransferRegistries.registerAdapter()` 注册。
- **新增传输事件系统**：新增 `PreTransferEvent`（可取消）和 `PostTransferEvent`，供第三方模组 hook 传输行为。
- **优化冷却管理器索引**：在 `CooldownManager` 中新增二级索引 `sourceKeyIndex`，支持 O(K) 移除指定 sourceKey 的所有冷却条目，提升清理性能。
- **优化容器配置缓存**：在 `ContainerConfig` 中新增 `cachedActualInterval` 字段，预计算冷却间隔，避免每次传输时重复计算。
- **优化过滤评估器缓存**：在 `FilterEvaluator` 中引入 `CompiledItemFilters` 和 `CompiledFluidFilters` 编译缓存，按 `FilterData` 实例缓存编译结果，提升过滤检查性能。
- **重构链接管理器职责分离**：将 `LinkManager` 中的脏数据追踪、保存调度、延迟网络同步逻辑分别拆分为 `LinkDirtyTracker`、`LinkSaveScheduler` 和 `PendingSyncBuffer` 独立组件，提升模块化和可维护性。
- **优化组服务与反向索引**：在 `GroupMemberService` 中新增 `nodeChannelIndex` 反向索引，支持 O(K) 注销节点的所有频道；在 `GlobalLogisticsManager` 中新增增量维护反向链接的 `addReverseLink` 和 `removeReverseLink` 方法。
- **优化传输日志结构**：将 `TransferLogManager` 拆分为 `TransferRecentLog`、`TransferStats`、`TransferRateCalculator` 和 `TransferLogManager` facade，提升线程安全性和模块清晰度。
- **新增集成指南文档**：新增 `docs/INTEGRATION.md`，详细说明如何通过 `LogisticsResource` 接口为第三方模组添加传输类型支持。
- **修复与调整**：修复蓝图粘贴时对无效方块的检查逻辑；优化 `NearestGroupSorter` 避免装箱开销；统一网络包中的版本号类型为 `long`；补充翻译键；更新 README 文档。

**行为对比**：
- 之前：传输系统基于 `TransferType` 记录和独立的 `ITransferHandler` 实现，代码重复且扩展复杂；冷却管理、过滤评估、链接管理、传输日志等组件职责混杂，性能有优化空间。
- 之后：所有传输类型通过统一的 `LogisticsResource` 接口接入，代码复用率高；各核心组件职责清晰分离，缓存机制增强，性能提升；提供完整的第三方集成指南和事件系统，扩展性显著增强。

+++ src/main/java/com/coobird/staticlogistics/client/data/ClientLinkData.java
@@ -15,6 +15,19 @@ import org.jetbrains.annotations.Nullable;
 import java.util.*;
 import java.util.concurrent.ConcurrentHashMap;
 
+/**
+ * 客户端物流数据存储 —— 接收服务端同步的面配置，供 GUI 和世界渲染使用。
+ *
+ * <p>职责：
+ * <ul>
+ *   <li>按维度存储所有面配置（由服务端通过 S2C 包同步）</li>
+ *   <li>维护已知组 ID、所有者名称、所有者 Profile 的缓存</li>
+ *   <li>维护服务端同步的空分组列表</li>
+ *   <li>提供查询接口供 GUI 和 {@link com.coobird.staticlogistics.client.render.LinkWorldRenderer} 使用</li>
+ * </ul>
+ *
+ * <p>线程安全：所有 Map 使用 ConcurrentHashMap，数据变更通过 {@code dataVersion} 计数器通知 GUI 刷新。
+ */
 @OnlyIn(Dist.CLIENT)
 public enum ClientLinkData {
     INSTANCE;
@@ -49,7 +62,7 @@ public enum ClientLinkData {
     /**
      * 更新面配置。服务端是唯一数据源，按序到达，客户端无条件接受。
      */
-    public void setFaceConfig(GlobalPos pos, Direction face, FaceConfigComposite config, int version) {
+    public void setFaceConfig(GlobalPos pos, Direction face, FaceConfigComposite config, long version) {
         long key = posToKey(pos.pos(), face);
         Map<Long, FaceConfigComposite> dimMap = getOrCreateDimMap(pos.dimension());
         if (config.isDefault()) {
